requirements.txt
django
python-dotenv
aiogram
psycopg2-binary
